Respect the presence of the packageCache
Our `package` event listener wasn’t working correctly when the event
was triggered via the `packageCache`, which `watchify` maintains. In
that scenario, there’s no `file` event following the `package` event.
Instead, we use the cached package result.

TODO: The order of the package events does not appear to be
deterministic (possibly in general, but especially) for cached
packages. This causes the deduping to potentially also be
non-deterministic as the algorithm considers the first module it comes
across to be the original. The watchify test added in the previous
commit reflects this, failing randomly.

[Fixes #8]